# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DFDR Swagger OpenAPI Documentation",
        "description": "Endpoint from DFDR Application",
        "contact": {
            "email": "admin@dfdr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/division": {
            "get": {
                "tags": [
                    "Division"
                ],
                "description": "View All",
                "operationId": "findAllDivision",
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Division"
                ],
                "description": "Add new",
                "operationId": "postDivision",
                "requestBody": {
                    "description": "Add new Division",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "division_name"
                                ],
                                "properties": {
                                    "division_name": {
                                        "description": "Division Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/division/{id}": {
            "get": {
                "tags": [
                    "Division"
                ],
                "description": "View By Id",
                "operationId": "showDivision",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Division",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            },
            "put": {
                "tags": [
                    "Division"
                ],
                "description": "Update Data",
                "operationId": "putDivision",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Division",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Add new Division",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "division_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Update Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Division"
                ],
                "description": "Delete Data",
                "operationId": "destroyDivision",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Division",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Delete Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/division/search/{name}": {
            "get": {
                "tags": [
                    "Division"
                ],
                "description": "Search Data",
                "operationId": "searchDivision",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Division",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Search Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer -token from login-)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}